requirements.txt
pydantic~=2.8.2
SQLAlchemy~=2.0.34
starlette~=0.38.6

fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pynacl==1.5.0